Development checking with masked arrays
This is an intermediate checkin, using numpy masked arrays as the data
structure, however it's clear that numba converts masked arrays to
plain numba arrays, so then argmin() doesn't work properly. The work
around is to hold the mask separately and pass this in as a separate,
argument so that the argmin() function can work on the correct values.

This means that the explicit masked arrays are not just an overhead,
so the next commits will drop them.