Switch away from masked arrays to manual mask
Masked arrays don't work very well as when jitted the wrong (unmasked)
version of argmin() is called, so previously merged pseudojets can be
mistakenly reconsidered.

Solution is to switch instead to normal numpy arrays, but also have a
mask array that can be used as needed to hide merged jet slots.

Added the antikt_distance array as we do need both - the nearest
neighbour is decided geomertically (dist), but the comparison metric is
decided also using the momentum of the jets (akt_dist).

Fixed a few bugs in which jets to updated after a merger and added
a missing abs() in the phi distance calculation.