Fixed accelerated tiled N^2 implementation
Critically changed the nn array from storing a tuple of indices
(which actually is then a 4D array) to a single flat index, which
is converted back to the index tuple as needed.

This allows the detection of invalid jets after an update to work
correctly, so keeps everything working properly - results are all now
in sync with fastjet outputs.