Optimise with a single jet scanner
Instead of scanning tile by tile, add a function
that only scans for a single jet's NNs. This is still
unidirectional, but it is a lot faster than doing
whole tile scans.

Correct a bug in nptiling where the nn matrix was being
set to bad values initially (hangover from the nn=tuple
old style schemea)