Enhance FastJet scripts
Add an option the allows different FastJet algorithms (inclusive-kt,
Cambridge/Achen, anti-kt) to be selected.

Improve the post-processing of FastJet outputs to not need output to be
trimmed and to allow the input and output files to be specified when
processing into JSON files.

Add a README to the FastJet directory describing compile and processing
steps.